extract ManageInventoryCardForm
